added lj potental to simulation subpackage, improved performance
